LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-select-element/customizable-select/base-appearance-inheritance-expected.txt



FAIL base appearance should not propagate to nested elements which support appearance. assert_equals: nested option should have auto appearance. expected "block" but got "flex"
